Move the full DAO Creation UI process out of page and into component
